Add datalayer interceptor configurations
Bambuser provides a few default interceptors for trying to detect add / remove to cart and product impression tracking events from the datalayer. These events are used to make an enhanced experience in the call, where the agent can see what items the shopper is looking at and adding to bag, without any advanced code integrations.
If the merchant has other data structures of these events that the GA4 standard, we now provide fields in the GTM template to configure where we can get the required values from.